Explicit cast to (int)
This is safe because the socketlist is resized to
    (size_t)s + 1
where s is an int.  Therefore
    for (i=0; i < socketlist_getsize(S); i++) {
will result in i being at most INT_MAX.

+++ events/events_network.c
@@ -273,7 +273,7 @@ events_network_select(struct timeval * tv)
 		for (i = 0; i < socketlist_getsize(S); i++) {
 			if (socketlist_get(S, i)->reader ||
 			    socketlist_get(S, i)->writer) {
-				fds[nfds].fd = i;
+				fds[nfds].fd = (int)i;
 				if (socketlist_get(S, i)->reader)
 					fds[nfds].events |= POLLIN;
 				if (socketlist_get(S, i)->writer)
